requirements.txt
ultralytics==8.1.0
opencv-python
numpy
torch
traci
matplotlib
tensorboard
PyYAML
